requirements.txt
requests>=2.31.0
pyyaml>=6.0
psycopg2-binary>=2.9.9
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
